+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 - Support for Wazuh 4.7.0
 
+### Changed
+
+- Changed Network ports table columns for Linux agents. [#5636](https://github.com/wazuh/wazuh-kibana-app/pull/5636)
+
 ### Fixed
 
 - Fixed problem with new or missing columns in agent table. [#5591](https://github.com/wazuh/wazuh-kibana-app/pull/5591)